appv2/requirements.txt
Pillow
Flask
Numpy
svgwrite
pymongo
pandas
unittest
